refactor(tui): split settings screen modules
Move settings rendering, input, keymap, types, tests, and tab field builders out of the monolithic settings screen module. Keep mod.rs as the coordinator for config sync and validation orchestration.

Closes #398
