+++ src/tui/screens/settings/tabs/flags.rs
@@ -1,0 +1,2 @@
+// The Flags tab is read-only: it has no editable widget fields and is rendered
+// directly by the settings draw module.
